すべてのepochのmodelを出力、logをcsvで出力するようにした
