+++ python/image/training/data_set.py
@@ -4,20 +4,7 @@ import glob
 import pandas as pd
 import sys
 from tensorflow.keras.preprocessing.image import ImageDataGenerator, load_img, img_to_array
-
-
-def clear_directory(path):
-    if os.path.exists(path):
-        # ディレクトリ内のファイルとサブディレクトリを削除
-        for filename in os.listdir(path):
-            file_path = os.path.join(path, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print(f'Failed to delete {file_path}. Reason: {e}')
+from util import clear_directory 
 
 
 # 生成枚数（１枚あたり）
