+++ python/image/training/training_log.csv
@@ -1,4 +1,4 @@
 epoch,accuracy,loss,val_accuracy,val_loss
-0,0.7333333492279053,0.9843536615371704,1.0,0.01987086795270443
-1,1.0,0.008264944888651371,0.9666666388511658,0.06809034198522568
-2,1.0,0.01038481667637825,0.6666666865348816,0.6463412046432495
+0,0.6416666507720947,0.8506534695625305,0.7333333492279053,0.7202374935150146
+1,1.0,0.009963983669877052,0.7666666507720947,0.9986143708229065
+2,1.0,0.0035664744209498167,0.7666666507720947,1.1397719383239746
